requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.4.2
geopy==2.3.0
python-dotenv==1.0.0 